organize components into subfolders
