+++ src/app/components/import/drop-image/drop-image.component.scss
