+++ src/app/components/import/drop-image/drop-image.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AppRepository, ImageFile } from '../../state/cutter.store';
+import { AppRepository, ImageFile } from '../../../state/cutter.store';
 
 @Component({
   selector: 'app-drop-image',
