+++ src/app/components/import/new-image/new-image.component.scss
