+++ src/app/components/layers/canvas-navigation/canvas-navigation.component.scss
