+++ src/app/components/layers/canvas-preview/canvas-preview.component.html
