+++ src/app/components/layers/layer-box/layer-box.component.scss
