+++ src/app/components/layers/layers-cut/layers-cut.component.scss
