+++ src/app/components/layers/layers/layers.component.scss
