+++ src/app/components/tabbar/tabs/tabs.component.scss
