+++ src/app/components/tabbar/tabs/tabs.component.spec.ts
