+++ src/app/components/toolbar/toolbar-cuts/toolbar-cuts.component.scss
